feat(GameInitializer, InitializedGame, GameStatePersister): Runner 책임 분리
