+++ src/main/java/InitializedGame.java
@@ -1,0 +1,4 @@
+import domain.GameDeadline;
+import domain.JanggiGame;
+
+public record InitializedGame(JanggiGame game, GameDeadline deadline) {}
